next-auth.d.ts => index.d.ts
