+++ apps/chat/index.d.ts
@@ -1,6 +1,23 @@
+import { DefaultSession } from 'next-auth';
+import { DefaultJWT } from 'next-auth/jwt';
+
 /* eslint-disable @typescript-eslint/no-explicit-any */
 declare module '*.svg' {
   const content: any;
   export const ReactComponent: any;
   export default content;
+}
+
+declare module 'next-auth' {
+  interface Session {
+    user: {
+      dial_roles?: string[];
+    } & DefaultSession['user'];
+  }
+}
+
+declare module 'next-auth/jwt' {
+  interface JWT extends DefaultJWT {
+    user: { dial_roles?: string[] };
+  }
 }
