--- apps/chat/next-auth.d.ts
@@ -1,16 +1,0 @@
-import { DefaultSession } from 'next-auth';
-import { DefaultJWT } from 'next-auth/jwt';
-
-declare module 'next-auth' {
-  interface Session {
-    user: {
-      dial_roles?: string[];
-    } & DefaultSession['user'];
-  }
-}
-
-declare module 'next-auth/jwt' {
-  interface JWT extends DefaultJWT {
-    user: { dial_roles?: string[] };
-  }
-}
